Add new export method
To be able to continue the modularity of the architecture, and also provide a base structure for any output format, I've refactored the old output_report.py file, that was responsible for parsing the nmap's output and also generating the LaTex report into two files. One, the parse_report.py file is responsible for only parsing the output from nmap's XML and generating a formatted JSON file inside the same directory, and a list of services in another JSON file, which will then be read by the second file, the export_latex.py file, that will continue to use the same logic for generating report. With this two components separeted, I was able to write another file, the export_neo4j.py file, that will get the same output format used to generate the LaTex report, and use it to feed a the Neo4j graph database, for interactive and permanent observability of services inside the network.