Add documentation and fix some bugs
Adding the proper documentation for usage about the Neo4j export format. Also add to the Dockerfile the new file names so the container can actually use the new scripts, and also fix a minor bug, caused by local testing, that to access the actual JSON report filename it just need -1 index in the array, not -2. Last, but not least, change the run.sh file so running without a $export variable actually runs the normal LaTeX export routine.